+++ eSlayerHelpers/include/eSlayerHelpers/eitem.h
@@ -20,10 +20,10 @@ struct ESLAYERHELPERS_API eItem {
 
     uint8_t fRequiredLevel = 0;
 
-    float fMinDmg = 0.f; // min weapon / shield / boots damage
-    float fMaxDmg = 0.f; // max weapon weapon / shield / boots damage
-    float fDefense = 0.f; // armor defense
-    float fBlockChance = 0.f; // shield block chance
+    uint16_t fMinDmg = 0; // min weapon / shield / boots damage
+    uint16_t fMaxDmg = 0; // max weapon weapon / shield / boots damage
+    uint16_t fDefense = 0; // armor defense
+    uint16_t fBlockChance = 0; // shield block chance
     std::vector<eModifier> fModifiers;
 
     void read(ePacket& p);
